cleaning up the test code a bit
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@2427 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/tests/test_feature/source/test_saveload.cpp
@@ -1296,7 +1296,7 @@ bool Test()
 		{
 			for( size_t b = 0; b < stream4.buffer.size(); ++b )
 				if( stream4.buffer[b] != stream2.buffer[b] )
-					PRINTF("streams differ on byte %d\n", b);
+					PRINTF("streams differ on byte %d\n", (int)b);
 		}
 		else
 			PRINTF("streams differ in size\n");
